Prepare for v2.0.0 of @actions/artifact
