--- packages/artifact/__tests__/ci-test-action/action.yml
@@ -1,5 +1,0 @@
-name: 'Set env variables'
-description: 'Sets certain env variables so that e2e artifact upload and download can be tested in a shell'
-runs:
-  using: 'node12'
-  main: 'index.js'
