+++ packages/artifact/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@actions/artifact",
-  "version": "1.1.1",
+  "version": "2.0.0",
   "preview": true,
   "description": "Actions artifact lib",
   "keywords": [
@@ -37,13 +37,10 @@
     "url": "https://github.com/actions/toolkit/issues"
   },
   "dependencies": {
-    "@actions/core": "^1.9.1",
-    "@actions/http-client": "^2.0.1",
-    "tmp": "^0.2.1",
-    "tmp-promise": "^3.0.2"
+    "@actions/core": "^1.10.0",
+    "@actions/http-client": "^2.1.0"
   },
   "devDependencies": {
-    "@types/tmp": "^0.2.1",
-    "typescript": "^3.8.3"
+    "typescript": "^3.9.10"
   }
 }
