--- packages/artifact/src/internal/download-options.ts
@@ -1,7 +1,0 @@
-export interface DownloadOptions {
-  /**
-   * Specifies if a folder is created for the artifact that is downloaded (contents downloaded into this folder),
-   * defaults to false if not specified
-   * */
-  createArtifactFolder?: boolean
-}
