--- packages/artifact/src/internal/download-response.ts
@@ -1,11 +1,0 @@
-export interface DownloadResponse {
-  /**
-   * The name of the artifact that was downloaded
-   */
-  artifactName: string
-
-  /**
-   * The full Path to where the artifact was downloaded
-   */
-  downloadPath: string
-}
